requirements.txt
pandas 
dvc 
mlflow==2.9.2
notebook
numpy
matplotlib
seaborn 
python-box==6.0.2
pyyaml
tqdm
ensure==1.0.2
joblib
types-pyyaml
scipy 
Flask
Flask-cors
-e .

